Update existing data
When we need to change values in a table, we use the UPDATE command. The general construction to update a row is:

UPDATE table
  SET column1 = value1,
      column2 = value2
  WHERE condition;

Note that UPDATE usually requires a WHERE clause to specify the row or rows to be updated. As with SELECT, if we don't specify a condition to restrict the rows, the command applies to all the rows in the table.