Integration of cyf_hotels DB  in pgsql with NodeJS
Loading data from a database with a GET endpoint

Retrieving individual records or

In the server.js:

- We first import the Pool class from the pg library, which is used to connect to a database
- We create a new connection (db) where we specify the credentials to connect to the cyf_hotel database
- We then create a new /customers endpoint where we use the method query() to send a SQL query to load all the customers from the table customers and return the results with result.rows. - - The query method returns a Promise: so we can access the returned rows using a .then block.
- We can write any valid SQL query that we learned in the query() method!
Note: There is no semicolon to end the SQL statement.

We started our server with node server.js and tried to reach the /customers endpoint to see the list of customers currently available in our cyf_hotels database. We can try to create/update/delete customers using psql to verify that our API always returns what is stored in our database.